requirements.txt
streamlit>=1.31.0
pillow>=10.0.0
numpy>=1.24.0
tifffile>=2023.7.0
